Move terminal actions to internal package
Currently actions are available as API types from "o.e.t.control"
bundle, but used only from internals of "o.e.t.v.ui" bundle.
So, action can be moved to internal package of of "o.e.t.v.ui" bundle to
reduce API surface.
