+++ terminal/bundles/org.eclipse.terminal.control/build.properties
@@ -1,5 +1,5 @@
 ###############################################################################
-# Copyright (c) 2003, 2018 Wind River Systems, Inc. and others.
+# Copyright (c) 2003, 2025 Wind River Systems, Inc. and others.
 # All rights reserved. This program and the accompanying materials
 # are made available under the terms of the Eclipse Public License 2.0
 # which accompanies this distribution, and is available at
@@ -16,6 +16,7 @@
 # Michael Scharf (Wind River) - split into core, view and connector plugins
 # Martin Oberhuber (Wind River) - fixed copyright headers and beautified
 # Anna Dushistova (MontaVista) - added icons
+# Alexander Fedorov (ArSysOp) - further evolution
 ###############################################################################
 bin.includes = .,\
                META-INF/,\
@@ -24,7 +25,6 @@ bin.includes = .,\
                .options,\
                README.txt,\
                about.html,\
-               icons/,\
                HelpContexts.xml
 
 source.. = src/
