+++ terminal/bundles/org.eclipse.terminal.view.ui/src/org/eclipse/terminal/view/ui/internal/actions/TerminalActionClearAll.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2004, 2018 Wind River Systems, Inc. and others.
+ * Copyright (c) 2004, 2025 Wind River Systems, Inc. and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License 2.0
  * which accompanies this distribution, and is available at
@@ -17,25 +17,26 @@
  * Martin Oberhuber (Wind River) - fixed copyright headers and beautified
  * Anna Dushistova (MontaVista) - [227537] moved actions from terminal.view to terminal plugin
  * Uwe Stieber (Wind River) - [260372] [terminal] Certain terminal actions are enabled if no target terminal control is available
+ * Alexander Fedorov (ArSysOp) - further evolution
  ********************************************************************************/
-package org.eclipse.terminal.control.actions;
+package org.eclipse.terminal.view.ui.internal.actions;
 
 import org.eclipse.terminal.control.ITerminalViewControl;
-import org.eclipse.terminal.internal.control.impl.ImageConsts;
+import org.eclipse.terminal.view.ui.internal.ImageConsts;
 
 public class TerminalActionClearAll extends AbstractTerminalAction {
 	public TerminalActionClearAll() {
 		super(TerminalActionClearAll.class.getName());
 
-		setupAction(ActionMessages.CLEARALL, ActionMessages.CLEARALL, null, ImageConsts.IMAGE_ELCL_CLEAR_ALL,
-				ImageConsts.IMAGE_DLCL_CLEAR_ALL, false);
+		setupAction(ActionMessages.CLEARALL, ActionMessages.CLEARALL, null, ImageConsts.ACTION_ClearAll_enabled,
+				ImageConsts.ACTION_ClearAll_disabled, false);
 	}
 
 	public TerminalActionClearAll(ITerminalViewControl target) {
 		super(target, TerminalActionClearAll.class.getName());
 
-		setupAction(ActionMessages.CLEARALL, ActionMessages.CLEARALL, null, ImageConsts.IMAGE_ELCL_CLEAR_ALL,
-				ImageConsts.IMAGE_DLCL_CLEAR_ALL, false);
+		setupAction(ActionMessages.CLEARALL, ActionMessages.CLEARALL, null, ImageConsts.ACTION_ClearAll_enabled,
+				ImageConsts.ACTION_ClearAll_disabled, false);
 	}
 
 	@Override
