+++ terminal/bundles/org.eclipse.terminal.view.ui/src/org/eclipse/terminal/view/ui/internal/tabs/TabFolderMenuHandler.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2011, 2021 Wind River Systems, Inc. and others. All rights reserved.
+ * Copyright (c) 2011, 2025 Wind River Systems, Inc. and others. All rights reserved.
  * This program and the accompanying materials are made available under the terms
  * of the Eclipse Public License 2.0 which accompanies this distribution, and is
  * available at https://www.eclipse.org/legal/epl-2.0/
@@ -8,6 +8,7 @@
  *
  * Contributors:
  * Wind River Systems - initial API and implementation
+ * Alexander Fedorov (ArSysOp) - further evolution
  *******************************************************************************/
 package org.eclipse.terminal.view.ui.internal.tabs;
 
@@ -30,16 +31,16 @@ import org.eclipse.swt.dnd.Transfer;
 import org.eclipse.swt.widgets.Menu;
 import org.eclipse.terminal.connector.TerminalState;
 import org.eclipse.terminal.control.ITerminalViewControl;
-import org.eclipse.terminal.control.actions.AbstractTerminalAction;
-import org.eclipse.terminal.control.actions.TerminalActionClearAll;
-import org.eclipse.terminal.control.actions.TerminalActionCopy;
-import org.eclipse.terminal.control.actions.TerminalActionPaste;
-import org.eclipse.terminal.control.actions.TerminalActionSelectAll;
 import org.eclipse.terminal.view.core.ITerminalsConnectorConstants;
 import org.eclipse.terminal.view.ui.ITerminalsView;
+import org.eclipse.terminal.view.ui.internal.actions.AbstractTerminalAction;
 import org.eclipse.terminal.view.ui.internal.actions.InvertColorsAction;
 import org.eclipse.terminal.view.ui.internal.actions.RenameTerminalAction;
 import org.eclipse.terminal.view.ui.internal.actions.SelectEncodingAction;
+import org.eclipse.terminal.view.ui.internal.actions.TerminalActionClearAll;
+import org.eclipse.terminal.view.ui.internal.actions.TerminalActionCopy;
+import org.eclipse.terminal.view.ui.internal.actions.TerminalActionPaste;
+import org.eclipse.terminal.view.ui.internal.actions.TerminalActionSelectAll;
 import org.eclipse.ui.IWorkbenchActionConstants;
 
 /**
